New Z Carraige using rack and pinion for Z control.
Modify PenTool for new servo position and new Z Carraige into which a
pen can be mounted. Complete Z Carriage now moves up and down by sliding
on a vertically mounted M3 shaft, and is driven by sevo via rack/pinion
gears.

The pen is mounted directly in the Z Carriage, but for the moment is not
quick or easy to mount/remove. A better method should be designed to
easily mount/umount pens.

To create the gearing, the publicDomainGearV1.1 from
http://www.thingiverse.com/thing:5505 is required.